chore: rename packages/ to handlers/
