+++ package/native-package/src/handlers/NetInfo.ts
