+++ package/native-package/src/handlers/deleteFile.ts
