+++ package/native-package/src/handlers/getLocalAssetUri.ts
