+++ package/native-package/src/handlers/shareImage.ts
