+++ package/native-package/src/handlers/takePhoto.ts
