+++ package/native-package/src/handlers/triggerHaptic.ts
